app/requirements.txt
astroid==2.2.5
backports.csv==1.0.7
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-crispy-forms==1.7.2
django-import-export==1.2.0
et-xmlfile==1.0.1
isort==4.3.18
jdcal==1.4.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pylint==2.3.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.3.5
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0